+++ packages/components/clickable/src/Clickable/index.ts
@@ -1,0 +1,2 @@
+export { default as Clickable, type ClickableProps } from "./Clickable";
+export { default as ClickableWrapper, type ClickableWrapperProps } from "./ClickableWrapper";