+++ packages/core/src/components/Dropdown/components/MultiValueContainer/MultiValueContainer.jsx
@@ -3,8 +3,7 @@ import { components } from "react-select";
 import cx from "classnames";
 import { useHiddenOptionsData } from "../../hooks/useHiddenOptionsData";
 import Counter from "../../../Counter/Counter";
-import Dialog from "../../../Dialog/Dialog";
-import DialogContentContainer from "../../../DialogContentContainer/DialogContentContainer";
+import { Dialog, DialogContentContainer } from "@vibe/dialog";
 import Chips from "../../../Chips/Chips";
 import classes from "./MultiValueContainer.module.scss";