+++ packages/core/src/components/Menu/MenuItem/components/MenuItemIcon/MenuItemIcon.tsx
@@ -1,12 +1,13 @@
 import React from "react";
 import { Icon } from "@vibe/icon";
-import Flex from "../../../../Flex/Flex";
+import { Flex } from "@vibe/layout";
 import cx from "classnames";
 import styles from "./MenuItemIcon.module.scss";
 import { type MenuItemIconProps } from "./MenuItemIcon.types";
 
 const MenuItemIcon = ({
   icon,
+  isRightIcon = false,
   type,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   label,
@@ -20,6 +21,7 @@ const MenuItemIcon = ({
     className={cx(
       styles.iconWrapper,
       {
+        [styles.rightIcon]: isRightIcon,
         [styles.disabled]: disabled,
         [styles.withBackgroundColor]: !!backgroundColor
       },