+++ packages/core/src/components/next/Dropdown/components/Trigger/MultiSelectTrigger.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Flex } from "../../../../Flex";
+import { Flex } from "@vibe/layout";
 import MultiSelectedValues from "../MultiSelectedValues/MultiSelectedValues";
 import DropdownInput from "./DropdownInput";
 import { useDropdownContext } from "../../context/DropdownContext";
@@ -23,7 +23,8 @@ const MultiSelectTrigger = () => {
     getToggleButtonProps,
     label,
     getLabelProps,
-    ariaLabel
+    ariaLabel,
+    minVisibleCount
   } = useDropdownContext<BaseListItemData>();
 
   return (
@@ -52,6 +53,7 @@ const MultiSelectTrigger = () => {
                   contextOnOptionRemove?.(item);
                 }}
                 renderInput={() => <DropdownInput inputSize="small" />}
+                minVisibleCount={minVisibleCount}
               />
             ) : (
               <Flex gap="xs" wrap>