+++ packages/docs/.storybook/main.ts
@@ -3,7 +3,7 @@ import type { StorybookConfig } from "@storybook/react-vite";
 import remarkGfm from "remark-gfm";
 import fs from "fs";
 
-const componentsFolder = path.resolve(__dirname, "../../../components");
+const componentsFolder = path.resolve(__dirname, "../../components");
 const components = fs.readdirSync(componentsFolder).reduce((acc: Record<string, string>, component) => {
   const componentFolderPath = path.resolve(componentsFolder, component);
   if (fs.statSync(componentFolderPath).isDirectory()) {
@@ -52,7 +52,6 @@ export default {
     "../src/**/*.stories.@(js|jsx|ts|tsx)",
     "../../core/**/*.mdx",
     "../../core/**/*.stories.@(js|jsx|ts|tsx)",
-    "../../../components/**/*.stories.@(js|jsx|ts|tsx)",
     "../../storybook-blocks/**/*.mdx",
     "../../storybook-blocks/**/*.stories.@(js|jsx|ts|tsx)"
   ],
@@ -89,6 +88,7 @@ export default {
           "@vibe/core/interactionsTests": path.resolve(__dirname, "../../core/src/tests/interactions-utils.ts"),
           "@vibe/core/next": path.resolve(__dirname, "../../core/src/components/next.ts"),
           "@vibe/core": path.resolve(__dirname, "../../core/src/index.ts"),
+          "@vibe/base": path.resolve(__dirname, "../../base/src/index.ts"),
           ...components
         }
       },