+++ scripts/bundle-check/compare-bundles.js
@@ -43,13 +43,10 @@ Array.from(allComponents)
     const prEntry = prMap.get(componentName);
 
     let displayName;
-    if (componentName.startsWith("components/")) {
-      const packageName = componentName.split("/")[1];
+    if (componentName.startsWith("packages/components/")) {
+      const packageName = componentName.split("/")[2];
       displayName = `@vibe/${packageName}`;
-    } else if (
-      componentName.includes("/next/") ||
-      (componentName.includes("/Modal/") && !componentName.includes("/LegacyModal/"))
-    ) {
+    } else if (componentName.includes("/next/")) {
       const baseName = path.basename(componentName, path.extname(componentName));
       displayName = `${baseName} (Next)`;
     } else {