+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.2](https://github.com/mondaycom/vibe/compare/@vibe/icons@1.11.1...@vibe/icons@1.11.2) (2025-11-26)
+
+**Note:** Version bump only for package @vibe/icons
+
+
+
+
+
+## [1.11.1](https://github.com/mondaycom/vibe/compare/@vibe/icons@1.11.0...@vibe/icons@1.11.1) (2025-11-06)
+
+**Note:** Version bump only for package @vibe/icons
+
+
+
+
+
 # [1.11.0](https://github.com/mondaycom/vibe/compare/@vibe/icons@1.10.0...@vibe/icons@1.11.0) (2025-09-04)